Move unknown version up in hierarchy
to avoid dependencies between converters.
