+++ chemclipse/plugins/org.eclipse.chemclipse.msd.converter.supplier.mmass/META-INF/MANIFEST.MF
@@ -15,8 +15,7 @@ Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.chemclipse.converter;bundle-version="0.8.0",
  org.eclipse.chemclipse.model;bundle-version="0.8.0",
  org.eclipse.chemclipse.support;bundle-version="0.8.0",
- org.apache.commons.commons-codec;bundle-version="1.16.0"
+ org.apache.commons.commons-codec;bundle-version="1.16.0",
+ org.apache.commons.lang3;bundle-version="3.13.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Bundle-ActivationPolicy: lazy
-Import-Package: org.apache.commons.lang3.tuple;version="3.1.0",
- org.eclipse.chemclipse.xxd.converter.supplier.io.exception
