+++ chemclipse/plugins/org.eclipse.chemclipse.msd.converter.supplier.mzml/META-INF/MANIFEST.MF
@@ -15,15 +15,14 @@ Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.chemclipse.converter;bundle-version="0.8.0",
  org.eclipse.chemclipse.model;bundle-version="0.8.0",
  org.eclipse.chemclipse.support;bundle-version="0.8.0",
- org.apache.commons.commons-codec;bundle-version="1.15.0"
+ org.apache.commons.commons-codec;bundle-version="1.15.0",
+ org.eclipse.chemclipse.xxd.converter.supplier.mzml,
+ org.apache.commons.lang3;bundle-version="3.13.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Bundle-ActivationPolicy: lazy
 Import-Package: jakarta.xml.bind,
  jakarta.xml.bind.annotation,
- jakarta.xml.bind.annotation.adapters,
- org.apache.commons.lang3.tuple;version="3.1.0",
- org.eclipse.chemclipse.xxd.converter.supplier.io.exception
+ jakarta.xml.bind.annotation.adapters
 Export-Package: org.eclipse.chemclipse.msd.converter.supplier.mzml.converter.io,
  org.eclipse.chemclipse.msd.converter.supplier.mzml.internal.converter,
- org.eclipse.chemclipse.msd.converter.supplier.mzml.internal.v110.model,
- org.eclipse.chemclipse.msd.converter.supplier.mzml.preferences
+ org.eclipse.chemclipse.msd.converter.supplier.mzml.internal.v110.model
