+++ chemclipse/plugins/org.eclipse.chemclipse.pcr.converter.supplier.rdml/META-INF/MANIFEST.MF
@@ -13,8 +13,7 @@ Require-Bundle: org.eclipse.core.runtime,
  org.eclipse.chemclipse.pcr.model;bundle-version="0.8.0",
  org.eclipse.chemclipse.logging;bundle-version="0.8.0",
  org.eclipse.chemclipse.processing;bundle-version="0.8.0",
- org.eclipse.chemclipse.support;bundle-version="0.8.0",
- org.eclipse.chemclipse.xxd.converter.supplier.ocx;bundle-version="0.9.0"
+ org.eclipse.chemclipse.support;bundle-version="0.8.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-17
 Bundle-ActivationPolicy: lazy
 Import-Package: jakarta.xml.bind,
